requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
crispy-bootstrap5==2023.10 ; python_version >= "3.11" and python_version < "4.0"
django-admin-interface==0.28.5 ; python_version >= "3.11" and python_version < "4.0"
django-colorfield==0.11.0 ; python_version >= "3.11" and python_version < "4.0"
django-crispy-forms==2.1 ; python_version >= "3.11" and python_version < "4.0"
django==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
drf-yasg==1.21.7 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
markdown==3.5.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.11" and python_version < "4.0"
python-slugify==8.0.3 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.4 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
sentry-sdk[django]==1.40.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
